requirements.txt
flask
flask_cors
flask_restful
flask_migrate
flask_sqlalchemy
firebase_admin
apscheduler
pgvector